chore(ui): New location for cn definition file to get a better organized package
